requirements.txt
flask
flask-session
argon2-cffi
flask-wtf
flask-limiter
flask-mail
--- IGNORE ---
python-dotenv